refactor(RacingGame.java): RacingGame의 상위 패키지 변경
현재 RacingGame은 레이싱 게임에 대한 책임만 지고 있고 view과 상호작용을
하지 않으므로 컨트롤로라 볼 수 없다. 따라서 RacingGame의 상위 패키지를
racingGame.controller에서 racingGame.models로 이동 후 Racing이라는
컨트롤러를 생성해 Application 클래스의 메인에 존재했던, 컨트롤러에
적합한 로직을 Racing으로 이동.

BREAKING CHANGE: RacingGame 상위 패키지 수정, Racing이라는 새로운
컨트롤러 생성

RacingGame:
as-is:
package: racingcar.controller

to-be:
package: racingcar.models

Racing:
as-is:
Application 클래스 main 클래스 메서드 로직:
public class Application {

    public static void main(String[] args) {
        final String carNames = Input.inputValidNames();
        final int repeats = Input.inputValidRepeats();

        final RacingGame racingGame = new RacingGame(carNames, repeats);
        Output.printResultMessage();
        while (!racingGame.isEnd()) {
            racingGame.race();
            Output.printTurnResult(racingGame.getThisTurnResult());
            racingGame.endThisTurn();
        }
        Output.printWinner(racingGame.getWinnersName());
    }
}

to-be:
public class Racing {
    private RacingGame racingGame;

    public void startRacing() {
        final String carNames = Input.inputValidNames();
        final int repeats = Input.inputValidRepeats();
        racingGame = new RacingGame(carNames, repeats);

        Output.printResultMessage();
        while (!racingGame.isEnd()) {
            racingGame.race();
            Output.printTurnResult(racingGame.getThisTurnResult());
            racingGame.endThisTurn();
        }
        Output.printWinner(racingGame.getWinnersName());
    }
}

public class Application {

    public static void main(String[] args) {
        Racing racing = new Racing();
        racing.startRacing();
    }
}
